+++ spring/src/main/java/me/laiyijie/job/swagger/model/JobErrorLog.java
@@ -15,7 +15,7 @@ public class JobErrorLog  implements Serializable {
   private static final long serialVersionUID = 1L;
 
   @JsonProperty("id")
-  private Integer id = null;
+  private Long id = null;
 
   @JsonProperty("workflowId")
   private Integer workflowId = null;
@@ -35,7 +35,7 @@ public class JobErrorLog  implements Serializable {
   @JsonProperty("content")
   private String content = null;
 
-  public JobErrorLog id(Integer id) {
+  public JobErrorLog id(Long id) {
     this.id = id;
     return this;
   }
@@ -45,11 +45,11 @@ public class JobErrorLog  implements Serializable {
    * @return id
   **/
   @ApiModelProperty(value = "")
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
